integrated figfile contents into analyzer.py
